requirements.txt
requests
google-generativeai
flask
python-dotenv
PyGithub
sendgrid
